+++ packages/dx-react-demos/src/bootstrap3/filtering/controlled-mode.jsx
