+++ packages/dx-react-demos/src/bootstrap3/filtering/filter-row.jsx
