+++ packages/dx-react-demos/src/bootstrap3/filtering/filter-row.test.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import Demo from './local-grouping-custom';
+import Demo from './filter-row';
 
-describe('BS3: integrated-grouping-custom', () => {
+describe('BS3 filtering: local filter row', () => {
   it('should work', () => {
     mount(<Demo />);
   });
