+++ packages/dx-react-demos/src/bootstrap3/grouping/controlled-mode.jsx
