+++ packages/dx-react-demos/src/bootstrap3/grouping/custom-advanced.jsx
