+++ packages/dx-react-demos/src/bootstrap3/grouping/custom.test.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import Demo from './local-grouping-static';
+import Demo from './custom';
 
-describe('BS3: integrated-grouping-static', () => {
+describe('BS3: integrated-grouping-custom', () => {
   it('should work', () => {
     mount(<Demo />);
   });
