+++ packages/dx-react-demos/src/bootstrap3/grouping/grouping-with-ui.jsx
