+++ packages/dx-react-demos/src/bootstrap3/grouping/grouping-with-ui.test.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import Demo from './local-grouping-with-ui';
+import Demo from './grouping-with-ui';
 
-describe('MUI: grouping by drag\'n\'drop', () => {
+describe('BS3: grouping by drag\'n\'drop', () => {
   it('should work', () => {
     mount(<Demo />);
   });
