+++ packages/dx-react-demos/src/bootstrap3/grouping/static.jsx
