+++ packages/dx-react-demos/src/bootstrap3/paging/controlled-mode.jsx
