+++ packages/dx-react-demos/src/bootstrap3/paging/uncontrolled-mode.jsx
