+++ packages/dx-react-demos/src/bootstrap3/sorting/controlled-mode.jsx
