+++ packages/dx-react-demos/src/bootstrap3/sorting/custom-sorting.jsx
