+++ packages/dx-react-demos/src/bootstrap3/sorting/custom-sorting.test.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import Demo from './local-filtering-controlled';
+import Demo from './custom-sorting';
 
-describe('MUI: local filtering controlled demo', () => {
+describe('BS3 sorting: custom local sorting demo', () => {
   it('should work', () => {
     mount(<Demo />);
   });
