+++ packages/dx-react-demos/src/bootstrap3/sorting/group-sorting.jsx
