+++ packages/dx-react-demos/src/material-ui/filtering/controlled-mode.jsx
