+++ packages/dx-react-demos/src/material-ui/filtering/filter-row.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import Demo from './local-filter-row';
+import Demo from './filter-row';
 
 describe('MUI: local filter row demo', () => {
   it('should work', () => {
