+++ packages/dx-react-demos/src/material-ui/grouping/controlled-mode.jsx
