+++ packages/dx-react-demos/src/material-ui/grouping/custom-advanced.jsx
