+++ packages/dx-react-demos/src/material-ui/grouping/grouping-with-ui.jsx
