+++ packages/dx-react-demos/src/material-ui/grouping/static.jsx
