+++ packages/dx-react-demos/src/material-ui/sorting/custom-sorting.jsx
